+++ circus-train-gcp/src/main/java/com/hotels/bdp/circustrain/gcp/GCPCredentialCopier.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,75 +17,60 @@ package com.hotels.bdp.circustrain.gcp;
 
 import static com.hotels.bdp.circustrain.gcp.GCPConstants.GCP_KEYFILE_CACHED_LOCATION;
 
-import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.UUID;
 
-import org.apache.commons.io.FileUtils;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.filecache.DistributedCache;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.mapreduce.filecache.DistributedCache;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
 
 import com.hotels.bdp.circustrain.api.CircusTrainException;
 
-class GCPCredentialCopier {
+@Component
+public class GCPCredentialCopier {
   private static final Logger LOG = LoggerFactory.getLogger(GCPCredentialCopier.class);
-  private static final String WORKING_DIRECTORY = System.getProperty("user.dir");
-  private static final String RANDOM_STRING = UUID.randomUUID().toString() + System.currentTimeMillis();
-  private static final String CACHED_CREDENTIAL_NAME = "/ct-gcp-key-" + RANDOM_STRING + ".json";
-  private static final String HDFS_GS_CREDENTIAL_DIRECTORY = "hdfs:///tmp/ct-gcp-" + RANDOM_STRING;
-  private static final String HDFS_GS_CREDENTIAL_ABSOLUTE_PATH = HDFS_GS_CREDENTIAL_DIRECTORY + CACHED_CREDENTIAL_NAME;
 
-  private final FileSystem fs;
-  private final Configuration conf;
-  private final String credentialProvider;
-
-  GCPCredentialCopier(FileSystem fs, Configuration conf, String credentialProvider) {
-    this.fs = fs;
-    this.conf = conf;
-    if (credentialProvider == null) {
-      throw new IllegalArgumentException("credentialProvider must be set");
-    }
-    this.credentialProvider = credentialProvider;
-  }
-
-  void copyCredentials() {
+  public void copyCredentials(
+      FileSystem fs,
+      Configuration conf,
+      GCPCredentialPathProvider credentialPathProvider,
+      DistributedFileSystemPathProvider dfsPathProvider) {
     try {
-      copyCredentialIntoWorkingDirectory();
-      copyCredentialIntoHdfs();
-      copyCredentialIntoDistributedCache();
+      Path source = credentialPathProvider.newPath();
+      Path destination = dfsPathProvider.newPath(conf);
+      copyCredentialIntoHdfs(fs, source, destination);
+      linkRelativePathInDistributedCache(conf, source, destination);
     } catch (IOException | URISyntaxException e) {
       throw new CircusTrainException(e);
     }
   }
 
-  private void copyCredentialIntoWorkingDirectory() throws IOException {
-    File source = new File(credentialProvider);
-    File destination = new File(WORKING_DIRECTORY + CACHED_CREDENTIAL_NAME);
-    destination.deleteOnExit();
-    LOG.debug("Copying credential into working directory {}", destination);
-    FileUtils.copyFile(source, destination);
-  }
-
-  private void copyCredentialIntoHdfs() throws IOException {
-    Path source = new Path(credentialProvider);
-    Path destination = new Path(HDFS_GS_CREDENTIAL_ABSOLUTE_PATH);
-    Path destinationFolder = new Path(HDFS_GS_CREDENTIAL_DIRECTORY);
+  private void copyCredentialIntoHdfs(FileSystem fs, Path source, Path destination) throws IOException {
+    /*
+     * The Google credentials file must be present in HDFS so that the DistCP map reduce job can access it upon
+     * replication.
+     */
+    Path destinationFolder = destination.getParent();
     fs.deleteOnExit(destinationFolder);
     LOG.debug("Copying credential into HDFS {}", destination);
     fs.copyFromLocalFile(source, destination);
   }
 
-  private void copyCredentialIntoDistributedCache() throws URISyntaxException {
-    LOG.debug("{} added to distributed cache with symlink {}", HDFS_GS_CREDENTIAL_DIRECTORY,
-        "." + CACHED_CREDENTIAL_NAME);
-    DistributedCache.addCacheFile(new URI(HDFS_GS_CREDENTIAL_ABSOLUTE_PATH), conf);
-    //The "." must be prepended for the symlink to be created correctly for reference in Map Reduce job
-    conf.set(GCP_KEYFILE_CACHED_LOCATION, "." + CACHED_CREDENTIAL_NAME);
+  private void linkRelativePathInDistributedCache(Configuration conf, Path source, Path destination)
+    throws URISyntaxException, IOException {
+    /*
+     * The "#" links the HDFS location for the key file to the local file system credential provider path so that the
+     * GoogleHadoopFileSystem can subsequently resolve it from a local file system uri despite it being in a Distributed
+     * file system when the DistCP job runs.
+     */
+    String cacheFileUri = destination.toString() + "#" + source;
+    DistributedCache.addCacheFile(new URI(cacheFileUri), conf);
+    LOG.info("mapreduce.job.cache.files : {}", conf.get("mapreduce.job.cache.files"));
+    conf.set(GCP_KEYFILE_CACHED_LOCATION, source.toString());
   }
 }