+++ circus-train-gcp/src/main/java/com/hotels/bdp/circustrain/gcp/context/GCPSecurity.java
@@ -24,11 +24,21 @@ public class GCPSecurity {
 
   private String credentialProvider;
 
+  private String distributedFileSystemWorkingDirectory;
+
   public String getCredentialProvider() {
     return credentialProvider;
   }
 
   public void setCredentialProvider(String credentialProvider) {
     this.credentialProvider = credentialProvider;
+  }
+
+  public String getDistributedFileSystemWorkingDirectory() {
+    return distributedFileSystemWorkingDirectory;
+  }
+
+  public void setDistributedFileSystemWorkingDirectory(String distributedFileSystemWorkingDirectory) {
+    this.distributedFileSystemWorkingDirectory = distributedFileSystemWorkingDirectory;
   }
 }